add improvements in design, testing

+++ src/com/parking/service/interfaces/FeeStrategy.java
@@ -1,7 +1,10 @@
 package com.parking.service.interfaces;
 
 import com.parking.model.Ticket;
+import java.math.BigDecimal;
 
 public interface FeeStrategy {
-    double calculate(Ticket ticket);
+
+    BigDecimal calculate(Ticket ticket);
+
 }
